v 1.10.2 new struct state

+++ include/sha1.h
@@ -2,6 +2,6 @@
 #define SHA1
 #include "util.h"
 unsigned char *getsha1f(FILE *fp);
-unsigned char *getsha1s(unsigned char *s, unsigned long long n);
+unsigned char *getsha1s(unsigned char *s, unsigned int n);
 
 #endif
